backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-jwt-extended==4.3.1
python-dotenv==0.19.0
tuya-iot-py-sdk==0.6.6